Updated the downsampling tool
The current downsampling tool is used for slideseq. While the read function can be adapted to different data types/read structures, the writing function is specific to the slideseq format. To tackle this issue, we modified the main function to adapt to the different read structures: C18X6C9M1X, 16C10M and 16C12M.